Moving clustering and graph code into src/conspiracies in appropriate packages
