+++ pyproject.toml
@@ -28,7 +28,7 @@ dependencies = [
     "confection>0.0.1,<0.1.0",
     # allenNLP packages
     "allennlp>=2.10.0,<2.11.0",
-    "allennlp_models>=2.10.0,<2.11.0",
+    "allennlp-models>=2.10.0,<2.11.0",
     # transformers
     "transformers>=4.11.3,<4.19.0",
     # base+ 
@@ -37,7 +37,13 @@ dependencies = [
     "pandas>=1.1.5,<1.5.0",
     "jsonlines>=3.1.0,<3.2.0",
     "openai",
-    "ndjson"
+    "ndjson",
+    "networkx",
+    "matplotlib",
+    "umap-learn",
+    "hdbscan",
+    "sentence-transformers",
+    "stop-words"
 ]
 
 [project.license]
